Moved dcm functions into dcm file
