+++ rhscripts/dcm.py
@@ -1,15 +1,18 @@
 #!/usr/bin/env python
 import os
-try:
-    import pydicom as dicom
-except ImportError:
-    import dicom
+import pydicom as dicom
 import configparser
 import glob
 from shutil import copyfile
 import datetime
+import numpy as np
 from pathlib import Path
-from typing import Optional
+from typing import Optional, Tuple, Callable, Dict
+from pydicom import dcmread
+import cv2
+import random
+import socket
+from rhscripts.version import __version__
 
 def __generate_uid_suffix() -> str:
     """ Generate and return a new UID """
@@ -18,14 +21,14 @@ def __generate_uid_suffix() -> str:
 def generate_StudyInstanceUID() -> str:
     """ Generate and return a new StudyInstanceUID """
     return '1.3.51.0.1.1.10.143.20.159.{}.7754590'.format(__generate_uid_suffix())
-    
+
 def generate_SeriesInstanceUID() -> str:
     """ Generate and return a new SeriesInstanceUID """
     return '1.3.12.2.1107.5.2.38.51014.{}11111.0.0.0'.format(__generate_uid_suffix())
 
 def generate_SOPInstanceUID(i: int) -> str:
     """ Generate and return a new SOPInstanceUID
-    
+
     Parameters
     ----------
     i : int
@@ -37,12 +40,12 @@ class Anonymize:
     """
     Anonymize script for DICOM file or folder containing dicom files
     Simply removes or replaces patient sensitive information.
-    
+
     from rhscripts.dcm import Anonymize
     anon = Anonymize()
     anon.anonymize_folder(dicom_original_folder,dicom_anonymized_folder)
     """
-    
+
     def __init__( self, verbose: bool=False, remove_private_tags: bool=False, sort_by_instance_number: bool=False ):
         """
         Parameters
@@ -57,11 +60,11 @@ class Anonymize:
         self.verbose = verbose
         self.remove_private_tags = remove_private_tags
         self.sort_by_instance_number = sort_by_instance_number
-    
-    def anonymize_dataset(self, dataset: dicom.dataset.Dataset, new_person_name: str="anonymous", 
+
+    def anonymize_dataset(self, dataset: dicom.dataset.Dataset, new_person_name: str="anonymous",
                   studyInstanceUID: str=None, seriesInstanceUID: str=None, replaceUIDs: bool=False) -> dicom.dataset.Dataset:
         """ Anonymize a single slice
-        
+
         Parameters
         ----------
         filename: str
@@ -76,57 +79,57 @@ class Anonymize:
             Overwrite instead of generating new. Used to make sure all slices have same SeriesInstanceUID.
         replaceUIDs: bool, optional
             Forces replacement of UIDs. Must be True when studyInstanceUID or seriesInstanceUID is set.
-            
+
         Return
         ------
         Anonymized dicom dataset
         """
-        
+
         new_person_name = "anonymous" if new_person_name == None else new_person_name
-    
+
         def __PN_callback(ds, data_element):
             """Called from the dataset "walk" recursive function for all data elements."""
             if data_element.VR == "PN":
                 data_element.value = new_person_name
-    
+
         # Remove patient name and any other person names
         dataset.walk(__PN_callback)
-    
-        # Remove data elements (should only do so if DICOM type 3 optional) 
+
+        # Remove data elements (should only do so if DICOM type 3 optional)
         for name in ['OtherPatientIDs', 'OtherPatientIDsSequence']:
             if name in dataset:
                 delattr(dataset, name)
-    
+
         # Same as above but for blanking data elements that are type 2.
         for name in ['PatientBirthDate','PatientAddress','PatientTelephoneNumbers']:
             if name in dataset:
                 dataset.data_element(name).value = ''
-    
+
         # Overwrite PatientID
         dataset.data_element('PatientID').value = new_person_name
-        
+
         # Overwrite AccessionNumber and StudyID
         if 'AccessionNumber' in dataset:
             dataset.data_element('AccessionNumber').value = new_person_name
         if 'StudyID' in dataset:
             dataset.data_element('StudyID').value = new_person_name
-    
+
         # Remove private tags
         if self.remove_private_tags:
             dataset.remove_private_tags()
-            
+
         # Replace InstanceUIDs
         if replaceUIDs:
             dataset.StudyInstanceUID = studyInstanceUID if studyInstanceUID is not None else generate_StudyInstanceUID()
             dataset.SeriesInstanceUID = seriesInstanceUID if seriesInstanceUID is not None else generate_SeriesInstanceUID()
             dataset.SOPInstanceUID = generate_SOPInstanceUID(dataset.InstanceNumber)
-    
+
         return dataset
-        
-    def anonymize_file( self, filename: str, output_filename: str, new_person_name: str="anonymous", 
+
+    def anonymize_file( self, filename: str, output_filename: str, new_person_name: str="anonymous",
                   studyInstanceUID: str=None, seriesInstanceUID: str=None, replaceUIDs: bool=False):
         """ Anonymize a single slice
-        
+
         Parameters
         ----------
         filename: str
@@ -136,24 +139,24 @@ class Anonymize:
         """
         # Load the current dicom file to 'anonymize'
         ds = dicom.read_file(filename)
-        
+
         if replaceUIDs:
             assert studyInstanceUID is not None # Must be set on folder level
             assert seriesInstanceUID is not None # Must be set on folder level
-            
+
         # Anonymize
         ds = self.anonymize_dataset( dataset=ds, new_person_name=new_person_name, studyInstanceUID=studyInstanceUID,
                                      seriesInstanceUID=seriesInstanceUID, replaceUIDs=replaceUIDs )
-        
+
         # Overwrite filename
         if self.sort_by_instance_number:
             output_filename = "dicom"+str(ds.InstanceNumber).zfill(4)+'.dcm'
-            
+
         # write the 'anonymized' DICOM out under the new filename
-        ds.save_as(output_filename) 
-        
-    def anonymize_folder(self,foldername: str,output_foldername: str, 
-                         new_person_name: str="anonymous", overwrite_ending: bool=False, 
+        ds.save_as(output_filename)
+
+    def anonymize_folder(self,foldername: str,output_foldername: str,
+                         new_person_name: str="anonymous", overwrite_ending: bool=False,
                          ending_suffix: str='.dcm', studyInstanceUID: str=None,
                          replaceUIDs: bool=False):
         """ Function to anonymize all files in the folder and subfolders.
@@ -181,22 +184,22 @@ class Anonymize:
             When output is not a directory.
 
         """
-        
+
         if os.path.exists(output_foldername):
             if not os.path.isdir(output_foldername):
                 raise IOError("Input is directory; output name exists but is not a directory")
         else: # out_dir does not exist; create it.
             os.makedirs(output_foldername)
-    
+
         if len(os.listdir(foldername)) > 9999:
             exit('Too many files in folder for script..')
-            
+
         # Check for replaceUIDs:
         if replaceUIDs:
-            if studyInstanceUID is None: 
+            if studyInstanceUID is None:
             	studyInstanceUID = generate_StudyInstanceUID()
             seriesInstanceUID = generate_SeriesInstanceUID()
-    
+
         for fid,filename in enumerate(os.listdir(foldername)):
             ending = '.dcm' if os.path.splitext(filename)[1]=='' else os.path.splitext(filename)[1]
             if overwrite_ending:
@@ -330,12 +333,12 @@ def send_data(folder, server=None, checkForEndings=True):
         Name of the server to send to
     checkForEndings : boolean, optional
         Check if folder contains any files with dicom endings
-    
+
     """
     from rhscripts.conversion import findExtension
     f = findExtension(folder)
     if not checkForEndings or isinstance(f,str):
-        
+
         # Setup
         if os.environ.get('CAAI') is not None:
             caai_dir = os.environ['CAAI']
@@ -349,36 +352,36 @@ def send_data(folder, server=None, checkForEndings=True):
         else:
             print('CAAI install path is not set in your environment')
             return
-        
+
         if not server:
             print('You need to select a server from config.ini')
             return
-        
+
         server = server.lower()
-        
+
         if not server in config:
             print('config.ini does not contain %s' % server)
             return
-        
-        
+
+
         cmd = 'storescu --scan-directories -aet %s -aec %s %s %s %s' % (
-                config['DEFAULT']['AET'], 
-                config[server]['AEC'], 
-                config[server]['addr'], 
-                config[server]['port'], 
+                config['DEFAULT']['AET'],
+                config[server]['AEC'],
+                config[server]['addr'],
+                config[server]['port'],
                 folder)
-        
+
         os.system(cmd)
-        
+
 def replace_container(in_folder: str, container: str, out_folder: str, SeriesNumber: int=None, SeriesDescription: str=None):
     """
-    
+
     Parameters
     ----------
     in_folder : str
         Path to folder with files containing PixelData that should be kept.
     container : str
-        Path to folder with files that should be used as container, getting only PixelData replaced 
+        Path to folder with files that should be used as container, getting only PixelData replaced
         (alongside UIDs and optionally SeriesNumber and SeriesDescription).
     out_folder : str
         Folder to store resulting output.
@@ -388,31 +391,436 @@ def replace_container(in_folder: str, container: str, out_folder: str, SeriesNum
         Overwrite the name of the series. The default is None.
 
     """
-    
+
     def sort_files(p):
         return {dicom.read_file(str(d)).InstanceNumber : d for d in Path(p).iterdir() }
 
     # Get dictionary with key=InstanceNumber val=Path-object for the file
     d_new = sort_files(in_folder)
     d_container = sort_files(container)
-    
-    # Create output folder   
+
+    # Create output folder
     Path(out_folder).mkdir(exist_ok=True,parents=True)
-    
+
     seriesInstanceUID = generate_SeriesInstanceUID()
-    
+
     # For each slice, 1-indexed due to InstanceNumber key
     for i in range(1,len(d_new)+1):
-        
+
         ds_container=dicom.read_file(str(d_container[i]))
         ds_new=dicom.read_file(str(d_new[i]))
-        
+
         ds_container.PixelData = ds_new.PixelData
-        
+
         ds_container.SeriesInstanceUID = seriesInstanceUID
         ds_container.SOPInstanceUID = generate_SOPInstanceUID( i )
-        
+
         if SeriesDescription is not None: ds_container.SeriesDescription = SeriesDescription
-        if SeriesNumber is not None: ds_container.SeriesNumber = SeriesNumber        
+        if SeriesNumber is not None: ds_container.SeriesNumber = SeriesNumber
 
         ds_container.save_as(f'{out_folder}/dicom_{i:04d}.dcm')
+
+def to_rtx(np_roi: np.ndarray,
+           dcmcontainer: str,
+           out_folder: str,
+           out_filename: str,
+           verbose: bool=False):
+
+    """Convert label numpy array to RT struct dicom file
+
+    Parameters
+    ----------
+    np_roi : np.ndarray
+        Numpy array in memory to be converted. Assume integer values of [0,1(,..)]
+    dcmcontainer : string
+        Path to dicom container to be used
+    out_folder : string
+        Name of the output folder
+    out_filename : string
+        Name of the output dicom file
+    verbose : boolean, optional
+        Verbosity of function
+    """
+
+    from pydicom.sequence import Sequence
+    from pydicom.dataset import Dataset, FileDataset
+    from decimal import Decimal, getcontext
+    getcontext().prec = 7
+
+    # Create affine transformation matrix:
+    def get_affine_transform(first_scan_path,last_scan_path,series_length):
+        # Load headers.
+        dicom_header_first = dicom.dcmread(first_scan_path)
+        dicom_header_last = dicom.dcmread(last_scan_path)
+
+        # Extract header info.
+        IOP = dicom_header_first.ImageOrientationPatient # Orientation is the same for all slices.
+        IOP = [float(i) for i in IOP]
+
+        IPP = dicom_header_first.ImagePositionPatient # Position of first slice.
+        IPP = [float(i) for i in IPP]
+
+        IPP2 = dicom_header_last.ImagePositionPatient # Position of last slice.
+        IPP2 = [float(i) for i in IPP2]
+
+        PS = dicom_header_first.PixelSpacing # Pixel Spacing is the same for all slices.
+        PS = [float(i) for i in PS]
+
+        # Creat empty matrix M.
+        M = np.zeros([4,4])
+
+        # Fill first column. Direction and Spacing for X coordinate.
+        M[0,0]=IOP[0]*PS[0]
+        M[1,0]=IOP[1]*PS[0]
+        M[2,0]=IOP[2]*PS[0]
+
+        # Fill second column. Direction and Spacing for Y coordinate.
+        M[0,1]=IOP[3]*PS[1]
+        M[1,1]=IOP[4]*PS[1]
+        M[2,1]=IOP[5]*PS[1]
+
+        # Fill third column. Direction and Spacing for Z coordinate.
+        M[0,2]=(IPP[0]-IPP2[0])/(1-series_length)
+        M[1,2]=(IPP[1]-IPP2[1])/(1-series_length)
+        M[2,2]=(IPP[2]-IPP2[2])/(1-series_length)
+
+        # Fill fourth column. Coordinate shift.
+        M[0,3]=IPP[0]
+        M[1,3]=IPP[1]
+        M[2,3]=IPP[2]
+        M[3,3]=1
+
+        return M
+
+    # Mask --> Polyline transform.
+    def get_polylines(mask,affine_transform_matrix,series_length):
+        # Create empty list to fill with contours.
+        polylines = []
+        for i in range(series_length): # Create a list space for each input slice. There are probably better ways to do this!
+            polylines.append([])
+
+        # Extract contours from mask and save as polylines.
+        for i in range(series_length):
+            contours, hierarchy = cv2.findContours(mask[:,:,i], cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE) # Convert masks to contours.
+            for contour in contours: # Loop through contours.
+                contour_list = np.ndarray.tolist(contour) # Convert numpy array to list.
+                polyline = list() # Create empty polyline container.
+                for triplet in contour_list: # Loop through contour coordinate triplets.
+                    triplet = triplet[0] # Modify list structure.
+                    triplet.append(i+1) # Add Z dimension.
+                    triplet.append(1) # Add fourth dimension for matrix multiplication.
+                    triplet = list(np.matmul(affine_transform_matrix,triplet)) # Multiply coordinate with affine transform matrix.
+                    triplet.pop() # Remove fourth dimension.
+                    for coord in triplet: # Loop through single coordinates.
+                        polyline.append(str(coord)) # Convert each number to string format for dicom header.
+                polylines[i].append(polyline) # Save polyline.
+        return polylines
+
+
+    # Define filter function for incoming image series:
+    def checkEqual(lst):
+       return lst[1:] == lst[:-1]
+
+    # Define function for expanding point and line delineations:
+    def expand_polyline(poly_input,M):
+
+        input_length = len(poly_input)
+        x_shift = Decimal((M[0][0])/2)
+        y_shift = Decimal((M[1][1])/2)
+
+        poly_input = [Decimal(item) for item in poly_input]
+
+        expanded_poly = list()
+
+        # Expand point delineations:
+        if input_length == 3:
+            expanded_poly.append([str(poly_input[0]+x_shift),str(poly_input[1]+y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]+x_shift),str(poly_input[1]-y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]-x_shift),str(poly_input[1]+y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]-x_shift),str(poly_input[1]-y_shift),str(poly_input[2])])
+
+            expanded_poly = [list(i) for i in set(map(tuple, expanded_poly))] # Remove duplicated triplets.
+
+        # Expand line delineations:
+        if input_length == 6:
+            expanded_poly.append([str(poly_input[0]+x_shift),str(poly_input[1]+y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]+x_shift),str(poly_input[1]-y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]-x_shift),str(poly_input[1]+y_shift),str(poly_input[2])])
+            expanded_poly.append([str(poly_input[0]-x_shift),str(poly_input[1]-y_shift),str(poly_input[2])])
+
+            expanded_poly.append([str(poly_input[3]+x_shift),str(poly_input[4]+y_shift),str(poly_input[5])])
+            expanded_poly.append([str(poly_input[3]+x_shift),str(poly_input[4]-y_shift),str(poly_input[5])])
+            expanded_poly.append([str(poly_input[3]-x_shift),str(poly_input[4]+y_shift),str(poly_input[5])])
+            expanded_poly.append([str(poly_input[3]-x_shift),str(poly_input[4]-y_shift),str(poly_input[5])])
+
+            expanded_poly = [list(i) for i in set(map(tuple, expanded_poly))] # Remove duplicated triplets.
+
+        polyline = list()
+        for item in expanded_poly:
+            polyline+=item
+
+        return polyline
+
+    # change path str to path object
+    if isinstance(dcmcontainer, str):
+        dcmcontainer = Path(dcmcontainer)
+
+    # gather the dicom slices from the container
+    dcm_list = [str(f) for f in dcmcontainer.iterdir() if not f.name.startswith('.') and f.is_file()]
+    ref_dict = dict() # Store dcm image SOP Instance UID's and file name for later use.
+    series_list = list() # For checking that all slices are from the same series.
+
+    # Fill out numpy matrix and reference dictionary.
+    for files in dcm_list:
+        dcm_slice = dcmread(files)
+        ref_dict[dcm_slice.InstanceNumber]=[dcm_slice.SOPInstanceUID,files]
+        series_list.append(dcm_slice.SeriesInstanceUID)
+
+    # Check that all dicom images in the folder belong to the same series.
+    if checkEqual(series_list) == False:
+        print("Error: All image slices must belong to the same series.")
+        sys.exit()
+
+    # Get first and last image from dicom series.
+    first_scan_path = ref_dict[1][1]
+    last_scan_path = ref_dict[len(dcm_list)][1]
+
+    # Read dicom header from first file.
+    dicom_header_first = dicom.dcmread(first_scan_path)
+
+    # Get current time.
+    time = datetime.datetime.now()
+
+    # Get image modality to determine SOP class UID for later.
+    image_modality = dicom_header_first.Modality
+    if image_modality == "CT":
+        SOP_class_UID = '1.2.840.10008.5.1.4.1.1.2'
+    elif image_modality == "MR":
+        SOP_class_UID = '1.2.840.10008.5.1.4.1.1.4'
+    elif image_modality == "PT":
+        #SOP_class_UID = '1.2.840.10008.5.1.4.1.1.20'
+        SOP_class_UID = '1.2.840.10008.5.1.4.1.1.128'
+
+    # Create new series instance UID.
+    newSIUID = generate_SeriesInstanceUID()
+
+    #%% Create RT file:
+
+    # Create dicom meta header.
+    file_meta = Dataset()
+
+    file_meta.MediaStorageSOPClassUID = '1.2.840.10008.5.1.4.1.1.481.3' # (0002,0002) Media Storage SOP Class UID
+    file_meta.MediaStorageSOPInstanceUID = newSIUID # (0002,0003) Media Storage SOP Instance UID
+    file_meta.TransferSyntaxUID = '1.2.840.10008.1.2' # (0002,0010) Transfer Syntax UID
+    file_meta.ImplementationClassUID = "1.2.826.0.1.3680043.8.691.0.21" # (0002,0012) Implementation Class UID
+    file_meta.ImplementationVersionName = '1.0' # (0002,0013) Implementation Version Name
+
+    # Create dicom dataset.
+    RTSTRUCT = FileDataset('RT',{},preamble=b"\0" * 128,file_meta=file_meta)
+
+    # Fill dicom header using the input scan series.
+    RTSTRUCT.InstanceCreationDate = time.strftime('%Y%m%d') # (0008,0012) Instance Creation Date (YYYYMMDD)
+    RTSTRUCT.InstanceCreationTime = time.strftime('%H%M%S') # (0008,0013) Instance Creation Time (HHMMSS)
+    RTSTRUCT.SOPClassUID = '1.2.840.10008.5.1.4.1.1.481.3' # (0008,0016) SOP Class UID
+    RTSTRUCT.SOPInstanceUID = newSIUID # (0008,0018) SOP Instance UID
+
+    # Patient Information.
+    RTSTRUCT.PatientName = dicom_header_first.PatientName # (0010,0010) Patient Name
+    RTSTRUCT.PatientID = dicom_header_first.PatientID # (0010,0020) Patient ID
+    RTSTRUCT.PatientBirthDate = dicom_header_first.PatientBirthDate # (0010,0030) Patient Birth Date
+    RTSTRUCT.PatientSex = dicom_header_first.PatientSex # (0010,0040) Patient Sex
+
+    # Study Information.
+    RTSTRUCT.StudyInstanceUID = dicom_header_first.StudyInstanceUID # (0020,000d) Study Instance UID
+    RTSTRUCT.StudyDate = dicom_header_first.StudyDate # (0008,0020) Study Date (YYYYMMDD)
+    RTSTRUCT.StudyTime = dicom_header_first.StudyTime # (0008,0030) Study Time (HHMMSS)
+    RTSTRUCT.OperatorsName = '' # (0008,1070) Operators Name
+    RTSTRUCT.ReferringPhysicianName = dicom_header_first.ReferringPhysicianName # (0008,0090) Referring Physician Name
+    RTSTRUCT.StudyID = dicom_header_first.StudyID # (0020,0010) Study ID
+    RTSTRUCT.AccessionNumber = dicom_header_first.AccessionNumber # (0008,0050) Accession Number
+    RTSTRUCT.StudyDescription = dicom_header_first.StudyDescription # (0008,1030) Study Description
+    if 'ReferencedStudySequence' in dicom_header_first:
+        RTSTRUCT.ReferencedStudySequence = dicom_header_first.ReferencedStudySequence # (0008,1110) Referenced Study Sequence
+
+    # Series Information.
+    RTSTRUCT.Modality = 'RTSTRUCT' # (0008,0060) Modality
+    RTSTRUCT.SeriesInstanceUID = newSIUID # (0020,000e) Series Instance UID
+    RTSTRUCT.SeriesNumber = str(200 + random.randint(1,200)) # (0020,0011) Series Number
+    RTSTRUCT.SeriesDate = time.strftime('%Y%m%d') # (0008,0021) Series Date
+    RTSTRUCT.SeriesTime = time.strftime('%H%M%S') # (0008,0031) Series Time
+    RTSTRUCT.SeriesDescription = out_filename # (0008,103e) Series Description
+
+    # Equipment Information.
+    RTSTRUCT.Manufacturer = 'RH-SCRIPTS v{}'.format(__version__) # (0008,0070) Manufacturer
+    RTSTRUCT.StationName = socket.gethostname() # (0008,1010) Station Name
+
+    # Structure Set.
+    RTSTRUCT.StructureSetLabel = 'ROI' # (3006,0002) Structure Set Label
+    RTSTRUCT.StructureSetName = '' # (3006,0004) Structure Set Name
+    RTSTRUCT.StructureSetDate = time.strftime('%Y%m%d') # (3006,0008) Structure Set Date (YYYYMMDD)
+    RTSTRUCT.StructureSetTime = time.strftime('%H%M%S') # (3006,0009) Structure Set Time (HHMMSS)
+
+    RTSTRUCT.ReferencedFrameOfReferenceSequence = Sequence([Dataset()]) # (3006,0010) Referenced Frame Of Reference Sequence
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].FrameOfReferenceUID = dicom_header_first.FrameOfReferenceUID # (0020,0052) Frame Of Reference UID
+
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence = Sequence([Dataset()]) # (3006,0012) RT Referenced Study Sequence
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].ReferencedSOPClassUID = '1.2.840.10008.3.1.2.3.2' # (0008,1150) Referenced SOP Class UID
+    if 'ReferencedStudySequence' in dicom_header_first:
+        RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].ReferencedSOPInstanceUID = dicom_header_first.ReferencedStudySequence[0].ReferencedSOPInstanceUID # (0008, 1155) Referenced SOP Instance UID
+
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].RTReferencedSeriesSequence = Sequence([Dataset()]) # (3006,0014) RT Referenced Series Sequence
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].RTReferencedSeriesSequence[0].SeriesInstanceUID = dicom_header_first.SeriesInstanceUID # (0020,000e) Series Instance UID
+
+    RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].RTReferencedSeriesSequence[0].ContourImageSequence = Sequence([Dataset()]) # (3006,0016) Contour Image Sequence
+    for i in range(len(ref_dict)):
+        contour = Dataset()
+        contour.ReferencedSOPClassUID = SOP_class_UID # (0008,1150) Referenced SOP Class UID
+        contour.ReferencedSOPInstanceUID = ref_dict[i+1][0] # (0008,1155) Referenced SOP Instance UID
+        if i == 0:
+            RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].RTReferencedSeriesSequence[0].ContourImageSequence[0] = contour
+        else:
+            RTSTRUCT.ReferencedFrameOfReferenceSequence[0].RTReferencedStudySequence[0].RTReferencedSeriesSequence[0].ContourImageSequence.append(contour)
+
+    RTSTRUCT.StructureSetROISequence = Sequence([Dataset()]) # (3006,0020)  Structure Set ROI Sequence
+    RTSTRUCT.ROIContourSequence = Sequence([Dataset()]) # (3006,0039)  ROI Contour Sequence
+
+    # Get Affine Transform Matrix.
+    M=get_affine_transform(first_scan_path,last_scan_path,len(dcm_list))
+
+    for i in range(np_roi.max()):
+
+        ROI_expanded = (np_roi == i+1).astype(int)
+        ROI_expanded = ROI_expanded.astype('uint8')
+
+        #polylines = get_polylines(ROI_expanded[:,:,:,i],M,len(dcm_list)) # Get polylines.
+        polylines = get_polylines(ROI_expanded,M,len(dcm_list)) # Get polylines.
+
+        roi_set = Dataset()
+        roi_set.ROINumber = str(i+1) # (3006,0022) ROI Number
+        roi_set.ReferencedFrameOfReferenceUID = dicom_header_first.FrameOfReferenceUID # (3006,0024) Referenced Frame of Reference UID
+        roi_set.ROIName = 'ROI_'+str(i+1) # (3006,0026) ROI Name
+        roi_set.ROIGenerationAlgorithm = 'AUTOMATIC' # (3006,0036) ROI Generation Algorithm
+        if i == 0:
+            RTSTRUCT.StructureSetROISequence[0] = roi_set
+        else:
+            RTSTRUCT.StructureSetROISequence.append(roi_set)
+
+        contour = Dataset()
+        contour.ROIDisplayColor = ['255', '0', '0'] # (3006,002a) ROI Display Color
+        contour.ReferencedROINumber = str(i+1) # (3006,0084) Referenced ROI Number
+        contour.ContourSequence = Sequence([Dataset()]) # (3006,0040)  Contour Sequence
+        if i == 0:
+            RTSTRUCT.ROIContourSequence[0] = contour
+        else:
+            RTSTRUCT.ROIContourSequence.append(contour)
+
+        contour_number = 1 # Initialize contour number.
+        for x in range(len(polylines)):
+            if polylines[x]:
+                for n in range(len(polylines[x])):
+                    poly = Dataset()
+                    poly.ContourImageSequence = Sequence([Dataset()]) # (3006,0016)  Contour Image Sequence
+                    poly.ContourImageSequence[0].ReferencedSOPClassUID = '1.2.840.10008.3.1.2.3.2' # (0008,1150) Referenced SOP Class UID
+                    poly.ContourImageSequence[0].ReferencedSOPInstanceUID = ref_dict[x+1][0] # (0008,1155) Referenced SOP Instance UID
+
+                    if len(polylines[x][n]) == 3 or len(polylines[x][n]) == 6: # Create additional triplets.
+                        polylines[x][n] = expand_polyline(polylines[x][n],M)
+
+                    poly.ContourGeometricType = 'CLOSED_PLANAR' # (3006,0042) Contour Geometric Type
+                    poly.NumberOfContourPoints = str(int(len(polylines[x][n])/3)) # (3006,0046) Number of Contour Points
+                    poly.ContourNumber = str(contour_number) # (3006,0048) Contour Number
+                    poly.ContourData = polylines[x][n] # (3006,0050) Contour Data
+
+                    if contour_number == 1:
+                        RTSTRUCT.ROIContourSequence[i].ContourSequence[0] = poly
+                    else:
+                        RTSTRUCT.ROIContourSequence[i].ContourSequence.append(poly)
+                    contour_number += 1 # Update contour number.
+
+    # Save RTSTRUCT file.
+    out_path = Path(out_folder).joinpath(out_filename+".dcm")
+    out_path.parent.mkdir(exist_ok=True,parents=True)
+    dicom.filewriter.write_file(str(out_path), RTSTRUCT, write_like_original=False)
+
+def read_rtx( dcmfile: str, img_size: Tuple[int,int,int],
+              fn_world_to_voxel: Callable, behavior: str='default',
+              verbose: bool=False ) -> Dict[int, Dict[str,np.ndarray]]:
+    """Read dcm file (RT struct) to dict of np.arrays - one for each ROI
+
+    Parameters
+    ----------
+    dcmfile : string
+        Path to the dicom file (RT struct)
+    img_size : (int,int,int)
+        Size of the reference volume that the RT struct was defined on.
+        Assuming (slice,width,height)
+    fn_world_to_voxel : string
+        Function to convert world coordinate to voxel coordinate
+    behavior : string
+        Chose how to convert to polygon. Options: default, mirada
+    verbose : boolean, optional
+        Default = False (if true, print info)
+    Examples
+    --------
+    >>> from rhscripts.conversion import read_rtx
+    >>> read_rtx('RTstruct.dcm',[127,344,344],volume.convertWorldToVoxel,verbose=False)
+    """
+    try:
+        RTSS = dicom.read_file(dcmfile)
+        ROIs = RTSS.ROIContourSequence
+        ROI_output = {}
+
+        if verbose:
+            print(RTSS.StructureSetROISequence[0].ROIName)
+            print("Found",len(ROIs),"ROIs")
+
+        for ROI_id,ROI in enumerate(ROIs):
+
+            data = np.zeros(img_size)
+            contour_sequences = ROI.ContourSequence
+
+            if verbose:
+                print(" --> Found",len(contour_sequences),"contour sequences for ROI:",RTSS.StructureSetROISequence[ROI_id].ROIName)
+
+            for contour in contour_sequences:
+                assert contour.ContourGeometricType == "CLOSED_PLANAR"
+
+                current_slice_i_print = 0
+
+                if verbose:
+                    print("\t",contour.ContourNumber,"contains",contour.NumberOfContourPoints)
+
+                world_coordinate_points = np.array(contour.ContourData)
+                world_coordinate_points = world_coordinate_points.reshape((contour.NumberOfContourPoints,3))
+                current_slice = np.zeros((img_size[1],img_size[2]))
+                voxel_coordinates_inplane = np.zeros((len(world_coordinate_points),2))
+                current_slice_i = 0
+                for wi,world in enumerate(world_coordinate_points):
+                    voxel = fn_world_to_voxel([-world[0],-world[1],world[2]])
+                    current_slice_i = voxel[0]
+                    voxel_coordinates_inplane[wi,:] = [voxel[2],voxel[1]]
+
+                current_slice_inner = np.zeros((img_size[1],img_size[2]),dtype=np.float32)
+                if behavior == 'default':
+                    converted_voxel_coordinates_inplane = np.array(np.round(voxel_coordinates_inplane),np.int32)
+                    cv2.fillPoly(current_slice_inner,pts=[converted_voxel_coordinates_inplane],color=1)
+                elif behavior == 'mirada':
+                    for x in range( img_size[1] ):
+                        for y in range( img_size[2] ):
+                            current_slice_inner[ x, y ] = cv2.pointPolygonTest(
+                                    np.array( [voxel_coordinates_inplane], dtype='float32' ),
+                                    (y+0.5,x+0.5),
+                                    False
+                            )
+                    current_slice_inner = (current_slice_inner>0).astype('uint8')
+                data[int(round(current_slice_i))] += current_slice_inner
+
+            # Remove even areas - implies a hole.
+            data[data % 2 == 0] = 0
+
+            ROI_output[ROI_id] = {'ROIname': RTSS.StructureSetROISequence[ROI_id].ROIName,
+                                  'data': data}
+        return ROI_output
+    except InvalidDicomError:
+        print("Could not read DICOM RTX file", dcmfile)
+        exit(-1)
